move connection header files
