+++ cpp/hal/connection_type/connection_aibom.h
