+++ cpp/hal/connection_type/connection_fullspec.h
