+++ cpp/hal/connection_type/connection_gamernium.h
