+++ cpp/hal/connection_type/connection_standard.h
